Create Teamviewer-ngrok-mirror.sh
This script will first connect to the remote computer using TeamViewer. Then, it will create a tunnel to the remote computer using ngrok. Finally, it will open a web browser and navigate to the ngrok URL. This will allow you to see the remote computer's desktop in your web browser.

To use this script, you will need to install TeamViewer and ngrok on your local computer. You will also need to know the name of the remote computer that you want to mirror