Update FCP-Zeroclick-Update.sh
These Bash script is a tool for remotely controlling an Android device using scrcpy. It works by generating a PDF payload that contains a scrcpy command. This payload is then encrypted and fragmented into small chunks. The fragmented payload is then sent to the Android device over the internet using ngrok.

When the Android device opens the PDF payload, the scrcpy command will be executed, giving you remote control over the device.

The script also prints the ngrok tunnel URL, which you can share with someone else so that they can also remotely control the device.