Update README.md
ANDROID FCP-Zeroclick
This bash script automates the process of remotely controlling an Android device using scrcpy and ngrok. It creates a fragmented PDF payload that contains the scrcpy command and encrypts it with a secret key. The script then starts ngrok to create a tunnel to the local port where the fragmented PDF payload is being served. Finally, the script opens the fragmented PDF file automatically in the background.

Usage
./FCP-Zeroclick-Update.sh
Requirements
ngrok
scrcpy
openvpn (optional)
Instructions
Replace the YOUR_SECRET_KEY placeholder in the script with your own secret key.
Replace the YOUR_VPN_CONFIG_FILE placeholder in the script with the path to your VPN configuration file, if you are using a VPN.
Make sure that the two PDF files that you want to combine are located in the paths specified by the pdf_file_1 and pdf_file_2 variables.
Run the script.
Example
./FCP-Zeroclick-Update.sh
This will create a fragmented PDF payload containing the scrcpy command and encrypt it with your secret key. The script will then start ngrok to create a tunnel to the local port where the fragmented PDF payload is being served. Finally, the script will open the fragmented PDF file automatically in the background.

Troubleshooting
If you are having trouble getting the script to work, please try the following:

Make sure that you have ngrok, scrcpy, and openvpn (optional) installed.
Make sure that the two PDF files that you want to combine are located in the paths specified by the pdf_file_1 and pdf_file_2 variables.
Make sure that your secret key is correct.
Try restarting your computer.