Create FCP-Zeroclick-Update.py
The Python script you provided is a remote control tool for Android devices that uses PDF files as a delivery mechanism. The script works by combining two PDF files into one, adding the scrcpy command to the PDF payload, encrypting the PDF payload, fragmenting the PDF payload, and then opening the PDF file on the Android device.

The scrcpy command is a free and open-source screen mirroring and remote control tool for Android devices. It allows you to mirror your Android device's screen to your computer and control your device remotely using your computer's keyboard and mouse.

To use the script, you will need to have the following installed:

Python 3
pdftk
openssl
scrcpy
ngrok
Once you have all of the required software installed, you can run the script by passing the paths to the two PDF files that you want to combine as arguments. The script will then generate a new PDF file that contains the scrcpy command. This PDF file will be encrypted and fragmented, and then opened on the Android device.

When the PDF file is opened on the Android device, the scrcpy command will be executed, and your computer will be able to mirror the Android device's screen and control the device remotely.

Here is a more detailed explanation of each step in the script:

Combine the two PDF files into one: This is done using the pdftk command.
Add the scrcpy command to the PDF payload: This is done using the pdftk command to insert the scrcpy command into the PDF payload.
Encrypt the PDF payload: This is done using the openssl command to encrypt the PDF payload with a secret key.
Fragment the PDF payload: This is done by sending the PDF payload to a local port in small chunks and then listening for the fragmented PDF payload on the local port and writing it to a temporary file.
Open the PDF file automatically in the background: This is done using the webbrowser module to open the PDF file.
Print the ngrok tunnel URL: This is done using the print() function to print the ngrok tunnel URL to the console.
Once the script has finished running, you will be able to control your Android device remotely by connecting to the ngrok tunnel URL from your computer.